Making 2 modules and importing them in main.js
