+++ public/index.html
@@ -13,7 +13,7 @@
   </head>
 
   <body>
-    <div style="display: none">
+    <div style="display: none;">
       <svg>
         <symbol
           id="wave"
@@ -146,6 +146,6 @@
         <button class="load-more" type="button">load more</button>
       </section>
     </main>
-    <script type="text/javascript" src="../src/main.js"></script>
+    <script type="module" src="../src/main.js"></script>
   </body>
 </html>
